requirements.txt
fastapi
uvicorn
python-multipart
pydub
replicate
python-dotenv
openai